Increased precision of latitude and longitude calculations.
git-svn-id: svn+ssh://garage/var/lib/gforge/svnroot/maemo-mapper/trunk@127 6c538b50-5814-0410-93ad-8bdf4c0149d1

+++ src/cmenu.c
@@ -46,7 +46,7 @@
 static void
 cmenu_show_latlon(gint unitx, gint unity)
 {
-  gfloat lat, lon;
+  gdouble lat, lon;
   gchar buffer[80], tmp1[LL_FMT_LEN], tmp2[LL_FMT_LEN];
   printf("%s()\n", __PRETTY_FUNCTION__);
 
@@ -69,7 +69,7 @@ static void
 cmenu_clip_latlon(gint unitx, gint unity)
 {
     gchar buffer[80];
-    gfloat lat, lon;
+    gdouble lat, lon;
     printf("%s()\n", __PRETTY_FUNCTION__);
 
     unit2latlon(unitx, unity, lat, lon);
@@ -88,7 +88,7 @@ cmenu_route_to(gint unitx, gint unity)
     gchar buffer[80];
     gchar strlat[32];
     gchar strlon[32];
-    gfloat lat, lon;
+    gdouble lat, lon;
     printf("%s()\n", __PRETTY_FUNCTION__);
 
     unit2latlon(unitx, unity, lat, lon);
@@ -106,7 +106,7 @@ static void
 cmenu_distance_to(gint unitx, gint unity)
 {
     gchar buffer[80];
-    gfloat lat, lon;
+    gdouble lat, lon;
     printf("%s()\n", __PRETTY_FUNCTION__);
 
     unit2latlon(unitx, unity, lat, lon);
